Move custom JSONEncoder to utils and make use of it for buffering.
