+++ setup.cfg
@@ -20,7 +20,7 @@ ignore = E123,E126,E203,E226,E241,E704,W503,W504
 match = ^((?!\.sync-zenodo-metadata|setup|benchmark|mpipool|connection|crypt|host|filesystems|indexing).)*\.py$
 match-dir = ^((?!\.|tests|configobj|db).)*$
 ignore-decorators = "deprecated"
-ignore = D105, D107, D203, D204, D213
+add-ignore = D105, D107, D203, D204, D213
 
 [mypy]
 ignore_missing_imports = True
